Refactor Item to be a Model
